test(infrastructure): move DB balance-validation tests to Postgres to fix CI host crash
CI's Build & Test runs dotnet test Receipts.slnx with XPlat Code Coverage, which
crashed the Infrastructure.Tests host (Total tests: Unknown -> MSB4181 -> exit 1)
even though every plain dotnet test run passed. The new InMemory unit tests for
CreateWithBalanceValidationAsync / UpdateWith... drove the service's
DbContext-per-operation fallback path (the InMemory provider has no row lock /
transaction), which is the code that misbehaved under coverage on the CI Linux runner.

Revert TransactionServiceTests to its mock-only form (the constructor now passes a
never-invoked mock IDbContextFactory plus the child mappers so it still compiles
against the new signature) and move the missing-receipt, soft-deleted (no-orphan),
balanced-persist, validator-rejection, and update-in-place coverage to a new
Infrastructure.IntegrationTests.TransactionBalanceValidationTests that runs against
real PostgreSQL alongside the existing concurrency test (RECEIPTS-763 / RECEIPTS-764).

No production code change.
