+++ tests/Infrastructure.Tests/Services/TransactionServiceTests.cs
@@ -1,14 +1,11 @@
 using Application.Models;
-using Domain;
 using Domain.Aggregates;
 using Domain.Core;
 using FluentAssertions;
-using FluentValidation;
 using Infrastructure.Entities.Core;
 using Infrastructure.Interfaces.Repositories;
 using Infrastructure.Mapping;
 using Infrastructure.Services;
-using Infrastructure.Tests.Repositories;
 using Microsoft.EntityFrameworkCore;
 using Moq;
 using SampleData.Domain.Core;
@@ -21,7 +18,6 @@ public class TransactionServiceTests
 	private readonly Mock<ITransactionRepository> _mockRepository;
 	private readonly TransactionMapper _mapper;
 	private readonly AccountMapper _accountMapper;
-	private readonly IDbContextFactory<ApplicationDbContext> _contextFactory;
 	private readonly TransactionService _service;
 
 	public TransactionServiceTests()
@@ -29,43 +25,23 @@ public class TransactionServiceTests
 		_mockRepository = new Mock<ITransactionRepository>();
 		_mapper = new TransactionMapper();
 		_accountMapper = new AccountMapper();
-		_contextFactory = DbContextHelpers.CreateInMemoryContextFactory();
+
+		// The balance-validation methods (CreateWithBalanceValidationAsync / UpdateWith...) require
+		// a DbContext factory and the child-entity mappers, but every test in this class exercises
+		// the repository-delegating members only, so the factory is never invoked. The DB-level
+		// balance-validation behaviour (row lock, existence guard, serialization) is covered against
+		// real PostgreSQL in Infrastructure.IntegrationTests.TransactionBalanceValidationTests —
+		// InMemory cannot model the row lock and running that path here crashed the test host under
+		// coverage on CI (RECEIPTS-764).
+		Mock<IDbContextFactory<ApplicationDbContext>> contextFactory = new();
 		_service = new TransactionService(
 			_mockRepository.Object,
 			_mapper,
 			_accountMapper,
-			_contextFactory,
+			contextFactory.Object,
 			new ReceiptMapper(),
 			new ReceiptItemMapper(),
 			new AdjustmentMapper());
-	}
-
-	// Balanced no-op validation delegate (the balance rule lives in the Application handlers).
-	private static readonly Action<ReceiptBalanceState> NoOpValidate = _ => { };
-
-	private async Task SeedReceiptWithItemAsync(Guid receiptId, bool softDeleted = false)
-	{
-		using ApplicationDbContext context = _contextFactory.CreateDbContext();
-		ReceiptEntity receipt = ReceiptEntityGenerator.Generate();
-		receipt.Id = receiptId;
-		await context.Receipts.AddAsync(receipt);
-		await context.ReceiptItems.AddAsync(ReceiptItemEntityGenerator.Generate(receiptId)); // TotalAmount $5, Tax $10 => ExpectedTotal $15
-		await context.SaveChangesAsync(CancellationToken.None);
-
-		if (softDeleted)
-		{
-			context.Receipts.Remove(receipt); // intercepted as a soft delete
-			await context.SaveChangesAsync(CancellationToken.None);
-		}
-	}
-
-	// Counts with IgnoreQueryFilters (no transaction is ever soft-deleted in these tests, so this
-	// is an exact row count) to avoid an InMemory quirk where the DeletedAt==null soft-delete
-	// filter is evaluated inconsistently right after a write.
-	private async Task<int> CountTransactionsAsync(Guid receiptId)
-	{
-		using ApplicationDbContext context = _contextFactory.CreateDbContext();
-		return await context.Transactions.IgnoreQueryFilters().CountAsync(t => t.ReceiptId == receiptId);
 	}
 
 	[Fact]
@@ -293,153 +269,5 @@ public class TransactionServiceTests
 
 		// Assert
 		result.Should().BeEmpty();
-	}
-
-	// ── CreateWithBalanceValidationAsync (RECEIPTS-763 / RECEIPTS-764) ──
-	// NOTE: the InMemory provider has no real transactions or row locks, so these tests verify
-	// the read-validate-write orchestration and the existence/soft-delete guard, NOT concurrent
-	// serialization. Cross-request serialization is exercised against real PostgreSQL in
-	// Infrastructure.IntegrationTests.TransactionBalanceConcurrencyTests.
-
-	[Fact]
-	public async Task CreateWithBalanceValidationAsync_ExistingReceipt_PersistsTransactionsAndReturnsThem()
-	{
-		// Arrange
-		Guid receiptId = Guid.NewGuid();
-		await SeedReceiptWithItemAsync(receiptId);
-		List<Transaction> input = [new(Guid.NewGuid(), Guid.NewGuid(), new Money(15), DateOnly.FromDateTime(DateTime.Now)) { AccountId = Guid.NewGuid() }];
-
-		// Act
-		List<Transaction> result = await _service.CreateWithBalanceValidationAsync(input, receiptId, NoOpValidate, CancellationToken.None);
-
-		// Assert
-		result.Should().HaveCount(1);
-		(await CountTransactionsAsync(receiptId)).Should().Be(1);
-
-		_contextFactory.ResetDatabase();
-	}
-
-	[Fact]
-	public async Task CreateWithBalanceValidationAsync_PassesFreshReceiptSnapshotToValidator()
-	{
-		// Arrange — the validator must see the receipt, its items, and its existing transactions
-		// as re-read INSIDE the guarded scope.
-		Guid receiptId = Guid.NewGuid();
-		await SeedReceiptWithItemAsync(receiptId);
-		using (ApplicationDbContext seed = _contextFactory.CreateDbContext())
-		{
-			await seed.Transactions.AddAsync(TransactionEntityGenerator.Generate(receiptId));
-			await seed.SaveChangesAsync(CancellationToken.None);
-		}
-
-		ReceiptBalanceState? captured = null;
-		List<Transaction> input = [new(Guid.NewGuid(), Guid.NewGuid(), new Money(1), DateOnly.FromDateTime(DateTime.Now)) { AccountId = Guid.NewGuid() }];
-
-		// Act
-		await _service.CreateWithBalanceValidationAsync(input, receiptId, state => captured = state, CancellationToken.None);
-
-		// Assert
-		captured.Should().NotBeNull();
-		captured!.Receipt.Id.Should().Be(receiptId);
-		captured.Items.Should().HaveCount(1);
-		captured.ExistingTransactions.Should().HaveCount(1);
-
-		_contextFactory.ResetDatabase();
-	}
-
-	[Fact]
-	public async Task CreateWithBalanceValidationAsync_MissingReceipt_ThrowsKeyNotFoundAndPersistsNothing()
-	{
-		// Arrange — no receipt seeded (RECEIPTS-763).
-		Guid receiptId = Guid.NewGuid();
-		List<Transaction> input = [new(Guid.NewGuid(), Guid.NewGuid(), new Money(15), DateOnly.FromDateTime(DateTime.Now)) { AccountId = Guid.NewGuid() }];
-
-		// Act
-		Func<Task> act = async () => await _service.CreateWithBalanceValidationAsync(input, receiptId, NoOpValidate, CancellationToken.None);
-
-		// Assert
-		await act.Should().ThrowAsync<KeyNotFoundException>();
-		(await CountTransactionsAsync(receiptId)).Should().Be(0);
-
-		_contextFactory.ResetDatabase();
-	}
-
-	[Fact]
-	public async Task CreateWithBalanceValidationAsync_SoftDeletedReceipt_ThrowsKeyNotFoundAndCreatesNoOrphan()
-	{
-		// Arrange — receipt exists in the table but is soft-deleted; the FK row still exists, so
-		// without the guard an ACTIVE transaction would be orphaned under a trashed receipt.
-		Guid receiptId = Guid.NewGuid();
-		await SeedReceiptWithItemAsync(receiptId, softDeleted: true);
-		List<Transaction> input = [new(Guid.NewGuid(), Guid.NewGuid(), new Money(15), DateOnly.FromDateTime(DateTime.Now)) { AccountId = Guid.NewGuid() }];
-
-		// Act
-		Func<Task> act = async () => await _service.CreateWithBalanceValidationAsync(input, receiptId, NoOpValidate, CancellationToken.None);
-
-		// Assert
-		await act.Should().ThrowAsync<KeyNotFoundException>();
-		(await CountTransactionsAsync(receiptId)).Should().Be(0);
-
-		_contextFactory.ResetDatabase();
-	}
-
-	[Fact]
-	public async Task CreateWithBalanceValidationAsync_ValidatorThrows_PersistsNothing()
-	{
-		// Arrange
-		Guid receiptId = Guid.NewGuid();
-		await SeedReceiptWithItemAsync(receiptId);
-		List<Transaction> input = [new(Guid.NewGuid(), Guid.NewGuid(), new Money(999), DateOnly.FromDateTime(DateTime.Now)) { AccountId = Guid.NewGuid() }];
-
-		// Act — validator rejects (mirrors the handler's balance-equation failure)
-		Func<Task> act = async () => await _service.CreateWithBalanceValidationAsync(
-			input, receiptId,
-			_ => throw new ValidationException("unbalanced"),
-			CancellationToken.None);
-
-		// Assert
-		await act.Should().ThrowAsync<ValidationException>();
-		(await CountTransactionsAsync(receiptId)).Should().Be(0);
-
-		_contextFactory.ResetDatabase();
-	}
-
-	[Fact]
-	public async Task UpdateWithBalanceValidationAsync_ExistingReceipt_UpdatesTransaction()
-	{
-		// Arrange
-		Guid receiptId = Guid.NewGuid();
-		await SeedReceiptWithItemAsync(receiptId);
-		TransactionEntity seeded = TransactionEntityGenerator.Generate(receiptId);
-		Guid seededAccountId = seeded.AccountId;
-		using (ApplicationDbContext seed = _contextFactory.CreateDbContext())
-		{
-			await seed.Transactions.AddAsync(seeded);
-			await seed.SaveChangesAsync(CancellationToken.None);
-		}
-
-		// Read the id actually assigned by the store (the InMemory key generator can reassign a
-		// ValueGeneratedOnAdd Guid on save; PostgreSQL keeps the client value). IgnoreAutoIncludes
-		// is required: the Account / Card navs are REQUIRED (non-nullable FK), so materializing the
-		// full entity would INNER JOIN and drop the row (no Account/Card is seeded here).
-		Guid seededId;
-		using (ApplicationDbContext read = _contextFactory.CreateDbContext())
-		{
-			seededId = (await read.Transactions.IgnoreQueryFilters().IgnoreAutoIncludes().ToListAsync()).Single().Id;
-		}
-
-		List<Transaction> update = [new(seededId, Guid.NewGuid(), new Money(42), DateOnly.FromDateTime(DateTime.Now)) { AccountId = seededAccountId }];
-
-		// Act
-		await _service.UpdateWithBalanceValidationAsync(update, receiptId, NoOpValidate, CancellationToken.None);
-
-		// Assert — the transaction was updated in place (amount changed) and NOT soft-deleted.
-		using ApplicationDbContext verify = _contextFactory.CreateDbContext();
-		TransactionEntity persisted = (await verify.Transactions.IgnoreQueryFilters().IgnoreAutoIncludes().ToListAsync())
-			.Single(t => t.Id == seededId);
-		persisted.Amount.Should().Be(42m);
-		persisted.DeletedAt.Should().BeNull();
-
-		_contextFactory.ResetDatabase();
 	}
 }
